Moved exceptions and validate into own file and improved exception hierarchy.
